address comments and fix errors
