+++ tox.ini
@@ -215,7 +215,11 @@ commands = python -m pip list --format=columns
 description = test error messages when importing unsupported modin or pandas versions
 basepython = python3.9
 setenv = TEST_INCORRECT_MODIN_PANDAS_VERSIONS=True
-commands = python -m pytest --noconftest tests/unit/modin/test_modin_pandas_version.py
+commands =
+    python -m pytest --noconftest tests/unit/modin/test_pandas_version.py
+    python -m pytest --noconftest tests/unit/modin/test_pandas_version_notebooks.py
+    python -m pytest --noconftest tests/unit/modin/test_modin_version.py
+    python -m pytest --noconftest tests/unit/modin/test_modin_version_notebooks.py
 
 [testenv:snowpark_pandas_old_np]
 description = test that Snowpark pandas works with numpy 1.26
